refactor: Use shorter name for the utility package
